Add files via upload
This is a Postman collection with a few examples for Workday SOAP and REST api calls. For them to work a "Integration System User" (ISU) and "API client for Integrations" need to be created in your Workday tenant with the correct permissions.

In the collection variables the client_id, client_secret and refresh_token need to be defined. 
As base_url variable the Workday community Gov shared demonstration tenant url is specified as example.

Before each request the pre-request Script will be executed which retrieves the Bearer access token and saves this as the AccessToken variable in the Postman collection which is then in turn be used for authentication of the request